move validator into seperate file and try and define pool type
